Generator - rozkład normalny, jednostajny hotfixy - polifazowy, merge sort

+++ src/main/java/algorithms/polyphasesort/PolyphaseSort.java
@@ -28,7 +28,7 @@ public class PolyphaseSort extends Sorter {
         output = sortedFile;
     }
 
-    public void sort(long blockSize) throws IOException {
+    public void sort(int blockSize) throws IOException {
         log.info("START - POLYPHASE SORT");
         int fakeSeriesNumber = 0;
         //podzial i sortowanie serii
